+++ app/src/main/java/io/github/kurramkurram/futaltacticalboard/db/SavedMovieListData.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "t_saved_movie_info")
 data class SavedMovieListData(
     @PrimaryKey(autoGenerate = true) val id: Int,
-    @ColumnInfo(name = "group_id") val groupId: Int,
+    @ColumnInfo(name = "group_id") val groupId: Long,
     @ColumnInfo(name = "title") val title: String,
     @ColumnInfo(name = "date") val date: String
 )
